Merged PR 116755: feat: prevent Story metadata override flag
# Short description
We added a boolean flag on the story module called `preventSettingMetadata` so, we need to handle this value to prevent the metadata override when the story module is inside a landing page.

## Note
Although the `preventSettingMetadata` has been defined as a `boolean`, the value returned by the  FrontendAPI is a string: since this is a well-known issue, at the moment I tried to compare the value as a string.

## List of changes proposed
- Add a flag to prevent `overrideDefaultMetadata` call inside the story module

## How to test the feature on localhost

- ENV: DEV
- URL: http://localhost:3000/test/react-poc/news-centre/latest-news

## Screenshots
A page containing an editorial list and a story:
![image (2).png](https://alm.deltatre.it/tfs/D3Alm/2feb7f47-11a3-4d82-a8ae-bbdb0e5ae438/_apis/git/repositories/9e4feb4e-e40e-4c06-84b6-657e6e8b667f/pullRequests/116755/attachments/image%20%282%29.png)

Result:
![image.png](https://alm.deltatre.it/tfs/D3Alm/2feb7f47-11a3-4d82-a8ae-bbdb0e5ae438/_apis/git/repositories/9e4feb4e-e40e-4c06-84b6-657e6e8b667f/pullRequests/116755/attachments/image.png)